Split ecoverse admin route
