+++ src/routing/admin/admin.route.tsx
@@ -4,7 +4,7 @@ import { managementData } from '../../components/Admin/managementData';
 import ManagementPageTemplatePage from '../../pages/Admin/ManagementPageTemplatePage';
 import { useTransactionScope } from '../../hooks';
 import { FourOuFour } from '../../pages';
-import { EcoverseListAdminRoute } from './ecoverse/ecoverse';
+import { EcoversesAdminRoute } from './ecoverse/ecoverses.admin.route';
 import { OrganizationsRoute } from './organization/organization';
 import { UsersRoute } from './users/users.route';
 import GlobalAuthorizationRoute from './GlobalAuthorizationRoute';
@@ -26,7 +26,7 @@ export const AdminRoute: FC = () => {
         <GlobalAuthorizationRoute paths={currentPaths} />
       </Route>
       <Route path={`${path}/hubs`}>
-        <EcoverseListAdminRoute paths={currentPaths} />
+        <EcoversesAdminRoute paths={currentPaths} />
       </Route>
       <Route path={`${path}/organizations`}>
         <OrganizationsRoute paths={currentPaths} />
